olsrd: nameservice: add postinst-hook, to upgrade plugin
olsrd-0.9.6 will come with an new versionnumber for the nameservice-plugin,
as of new parameter "filewrite_interval".
This commit installs a postinst-hook for mod-nameservice to update the
UCI-option library to the new value.

As this new plugin-parameter is compatible with the current behaviour, no
additional change to the current config is required.

At compile-time the value of $PLUGIN_CURRENT_VERSION in postinst is defined
from filename of plugin

Signed-off-by: Sven Roederer <freifunk@it-solutions.geroedel.de>