Fix NoDogSplash Startup Failure
Fix NoDogSplash Startup Failure if Device is Not Ready. 
The function wait_for_interface() was incorporated into the nodogsplash init script to prevent startup failure if the device for nodogsplash to bind to is not yet up when nodogsplash starts at boot time.
However if option network is defined in the configuration file, the function network_get_device() is used to determine the required device. 
Unfortunately if the device is not ready when network_get_device() is called, null is returned without error resulting in an empty temporary .config file being created  in /tmp/etc. NoDogsplash then fails.
The proposed updated config file defaults to setting option gatewayinterface 'br-lan'.