bird: remove incorrect build depends
Build depends must refer to source rather than binary package names
(readline and ncurses instead of libreadline and libncurses).

As libreadline and libncurses are also runtime dependencies of the
individual binary packages, build depends on the corresponding source
packages are implied and the incorrect build depends can simply be removed.

Signed-off-by: Matthias Schiffer <mschiffer@universe-factory.net>