miniupnpd: bump to 20180422
Refresh patches.
Drop 102-ipv6-ext-port.patch as this looks upstreamed in the pinhole
code to me.

Add support for runtime IGDv1 mode switch (default to IGDv2)

(not extensively) Tested-on: ar71xx Archer C7 v2 in IGDv1 compatibility
mode.  A variety of devices/applications appear to be able to create
mappings.

Signed-off-by: Kevin Darbyshire-Bryant <ldir@darbyshire-bryant.me.uk>