[WIP] babeld: expose add_xroute function via ubus
If we want to do a Layer3 handover, we need to redirect the traffic
to the new AP. However, this does not have to happen for long. A static
configuration would probably be too much.

This babel function will allow us to announce routes via ubus.

ubus call babeld add_xroute {
	"prefix": "2001::1",
	"plen": 128,
	"src_prefix": "0::"
	"src_plen": 0,
	"metric": 1,
	"ifindex": 1,
	"proto": 123
}

Signed-off-by: Nick Hainke <vincent@systemli.org>