opennds: avoid circular dependency
Have the package depend on @IPTABLES_NFTABLES, which is needed to be
able to select iptables-nft.

What you need to avoid is the selection of iptables-nft when
IPTABLES_NFTABLES is not selected, since iptables-nft depends on
xtables-nft, which depends on IPTABLES_NFTABLES.  If you have this
situation, when IPTABLES_INFTABLES is not selected, xtables-nft will not
be built, but opennds will.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>