ci: change default packages
Since this file was taken from packages repository, it tries to compile
packages from that repository as fallback if changes are not detected.

Let's change these packages to some, which are present in this
repository, so they will be compiled as test.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>
(cherry picked from commit c9a8f2d03226191111b95890302cb7e1ff18c11e)