treewide: remove AUTORELEASE
OpenWrt/packages removed AUTORELEASE treewide. Remove it also in the
routing feed.

This is just copied from [0] with modification to the sed cmd because it
was not working for the routing feed:

The following temporary change was made to the core:

diff --git a/rules.mk b/rules.mk
index 57d7995d4fa8..f16367de87a8 100644
--- a/rules.mk
+++ b/rules.mk
@@ -429,7 +429,7 @@ endef
 abi_version_str = $(subst -,,$(subst _,,$(subst .,,$(1))))

 COMMITCOUNT = $(if $(DUMP),0,$(call commitcount))
-AUTORELEASE = $(if $(DUMP),0,$(call commitcount,1))
+AUTORELEASE = $(if $(DUMP),0,$(shell sed -i "s/\$$(AUTORELEASE)/$(call commitcount,1)/" $(CURDIR)/Makefile))

 all:
 FORCE: ;

And this command used to fix affected packages:

for i in $(cd feeds/routing; git grep -l PKG_RELEASE:=.*AUTORELEASE | \
                              sed 's/\/Makefile$//';);
do
  make package/$i/download
done

[0] - https://github.com/openwrt/packages/commit/0c10c224be81f9221dabfab449855ab6718d2a0c

Signed-off-by: Nick Hainke <vincent@systemli.org>