olsrd: remove dependency on iptables virtual package
With the switch from OPKG to APK, we can't continue the
sloppy use of virtual packages. We now need to specify the
providing package. We don't want to force iptables-nft on
everyone though. The simplest solution is to drop the
dependency entirely - users will most likely already
depend on it separately.

For reference, the attempt to install the virtual package
results in the following error from APK:

ERROR: unable to select packages:
  ip6tables (virtual):
    note: please select one of the 'provided by'
          packages explicitly
    provided by: ip6tables-nft
                 ip6tables-zz-legacy
    required by: olsrd-2024.06.09~d72be9ad-r1[ip6tables]
  iptables (virtual):
    note: please select one of the 'provided by'
          packages explicitly
    provided by: iptables-nft iptables-zz-legacy
    required by: olsrd-2024.06.09~d72be9ad-r1[iptables]

Suggested-by: Packet Please <pktpls@systemli.org>
Link: https://github.com/openwrt/routing/pull/1085
[Due to the real name policy, list the author as me, with
permission granted by Packet Please.]
Signed-off-by: Nick Hainke <vincent@systemli.org>