bird2: update to 2.16.1
The bird package has been updated to include a patch that directly fixes the
alignment issue, which previously required the use of the -mno-unaligned-access
GCC flag on ARM processors. This patch, originally discussed on the BIRD mailing
list [0], modifies the net_addr structure to resolve the alignment problem
without needing the workaround.

By incorporating this patch, the alignment issue is addressed more efficiently,
eliminating the performance, code size, and hardware optimization drawbacks
associated with the GCC flag.

Release Notes:
http://trubka.network.cz/pipermail/bird-users/2025-January/018041.html

[0] - http://trubka.network.cz/pipermail/bird-users/2024-December/017957.html

Signed-off-by: Nick Hainke <vincent@systemli.org>